requirements.txt
PySide6>=6.4.0
PySide6-QtSvg
yt-dlp>=2023.10.13
requests>=2.28.1
Pillow>=9.3.0
pyinstaller>=5.13.0
